chore: Add Termly consent preferences and policies links (#3187)
* chore: Add Termly consent preferences and policies links

Added Termly consent preferences link and links to Privacy Policy, Cookie Disclaimer, Terms & Conditions, and Acceptable Use Policy in the HTML file. Also included a script for blocking resources from Termly.

* feat: Add consent preferences menu option

This commit adds a new menu option for "Consent Preferences" in the UI. When clicked, it triggers a click event on the consent link element and focuses on it. This allows users to easily access and manage their consent preferences.